Refactor + Root Transform T
